Use fixed keymaps for both minor modes
`local-set-key` docstring says: "The binding goes in the current buffer's local
map, which in most cases is shared with all other buffers in the same major
mode." It breaks our keymaps one way or the other as we open new Ruby files.

The only drawback I see here is that rspec-key-command-prefix can't be changed
once the mode file is loaded. If that's a desirable feature, we can add a
defcustom setter.