Fix yas-snippet-dirs to append rspec-snippet-dir
The first directory in yas-snippets-dirs is used as the default location
to save a new snippet to. The default documentation for running
rspec-install-snippets will add rspec-snippets-dir to the list very
late. In order to be respectful to the users defaults we should append
rspec-snippet-dir to the list instead of placing it at the beginning.