Fix bug in `rspec--include-fg-syntax-methods-p`
When called with `nil`, cl-case always return `nil`, even when there is an
explicit `nil` case.

For example,

```elisp
(setq myvar nil)
(cl-case myvar
  (one 1)
  (two 2)
  (nil 3)) # => nil, and not 3!
```

That is probably not what was meant by the author of
`rspec--include-fg-syntax-methods-p`. This commit changes the `nil`
case in `rspec--include-fg-syntax-methods-p` to `t`, so that it will
become the default case.