Add `rspec-verify-method`
This commit adds the function `rspec-verify-method` that behaves like
`rspec-verify-single`, but tries to guess which examples to run based on
the method at given point.

The implementation piggy-backs in the functionality provided by
`rspec-toggle-spec-and-target-find-example` for navigation between a
method and it's describe block (and vice-versa)

`rspec-verify-method` is intended to be used only on `code` buffers and
not on `spec` buffers. That's why the keybinding for
`rspec-verify-method` is added in the `rspec-verifiable-mode-keymap`