Look in rails helper for factory girl syntax methods.
More recent versions of rspec-rails create/recommend creating both a
`rails_helper.rb` and `spec_helper.rb` file, with the intention of
putting rails-specific configuration in the former. Given that a rails
project is set up in this way, including the syntax methods for factory
girl would be expected to go into the rails helper.

This change looks in the rails helper if it exists first before
searching the spec helper for the inclusion of this module.

https://github.com/pezra/rspec-mode/issues/117