Correctly match capybara-screenshot output
It appears upstream capybara-screenshot changed the output for reporting
screenshot failures:

https://github.com/mattheworiordan/capybara-screenshot/commit/4247c4296fc1a0cd8bf10a897f2511f0ed612435#diff-814c9a0aeab384689919406b7d82dabdL15

Updates the regex to match. It unfortunately doesn't entirely work,
because the link is of format
`file://$current_working_dir/path/to/screenshot{.png,.html}`. After
hitting enter emacs prompts with that as the default but doesn't seem to
find the resulting file.

At the very least this fixes the links to be detected correctly, not
sure on restoring the functionality to open on click.